feat(标注平台完成): 增加了任务用户表和未分配统计
1.增加任务用户查询
2.增加审核分配
3.增加未分配统计
4.尚未增加任务分配更加优化的逻辑

+++ ruoyi-admin/src/main/java/com/ruoyi/task/domain/TaskUserTaskAllocation.java
@@ -55,7 +55,7 @@ public class TaskUserTaskAllocation extends BaseEntity
 
 
     /** 召回数 */
-    private Double recallNum;
+    private Integer recallNum;
 
 
     private Long responsiblePersonId;
